Added services so that the controller can test what the actor should do
Added:
 Score level - service to what the actor should do at each level
 Test actor range - test if an actor has moved close enough to a target actor
 Test actor location - checks if the actor is near a location

+++ Plugins/AIPlugins/Memory/Source/Memory/Public/BehaviorAction.h
@@ -17,10 +17,10 @@ struct MEMORY_API FBehaviorAction
     int32 Level{ 0 };
 
     /**
-    * The priority of the action
+    * The priority of the action, the higher that this value is then the more likely it will be executed
     */
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Behavior")
-    float Priority{ 0.0f };
+    double Priority{ 0.0 };
 
     /**
     * Used to identify what this behavior action is
@@ -28,6 +28,10 @@ struct MEMORY_API FBehaviorAction
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Behavior")
     FString Note;
 
+    /**
+    * The value of the action, this would determine what action occurs e.g. 0 = Move, 1 = Attack.
+    * These values could come from an enum for clarity.
+    */
     UPROPERTY(EditAnywhere, BlueprintReadWrite, Category = "Behavior")
     int32 Value{ 0 };
 };
